// Front-end/Mobile Front-end/Admin-Front-end/types/message.d.ts
export interface IMessage {
    id: number;
    match_id: number;
    sender_id: number;
    content: string;
    message_type: "text" | "audio" | "video" | "image" | "file" | undefined;
    sent_at: string;
    read_at: string;
    deleted_at: string;
    is_pinned: boolean;
    pinned_at: string;
    reply_to_message_id: number;
    edited_at: string;
}